Move groupBy to libs/basics
