+++ apps/design/backend/package.json
@@ -46,7 +46,6 @@
     "@votingworks/utils": "workspace:*",
     "better-sqlite3": "^7.5.0",
     "debug": "^4.3.1",
-    "deep-eql": "^4.1.3",
     "dotenv": "^16.0.2",
     "dotenv-expand": "^9.0.0",
     "express": "^4.18.0",
@@ -63,7 +62,6 @@
     "@jest/types": "^29.6.1",
     "@types/better-sqlite3": "^7.4.3",
     "@types/debug": "^4.1.5",
-    "@types/deep-eql": "^4.0.0",
     "@types/express": "^4.17.13",
     "@types/fs-extra": "^9.0.6",
     "@types/jest": "^29.5.3",
