+++ libs/basics/package.json
@@ -31,9 +31,11 @@
     ]
   },
   "dependencies": {
+    "deep-eql": "^4.1.3",
     "util": "^0.12.4"
   },
   "devDependencies": {
+    "@types/deep-eql": "^4.0.0",
     "@types/jest": "^29.5.3",
     "@types/node": "16.18.23",
     "eslint-plugin-vx": "workspace:*",
