+++ libs/basics/src/index.ts
@@ -3,6 +3,7 @@ export * from './assert';
 export * as collections from './collections';
 export * from './deferred';
 export * from './duplicates';
+export * from './group_by';
 export * from './errors';
 export * from './find';
 export * from './iterators';
